requirements/base.txt
# Django
Django

# Models
#django-model-utils==2.4

# Rest apis
djangorestframework==3.3.2
#Markdown==2.6.5
#django-filter==0.11.0

# Static files
#whitenoise==2.0.6

# Media files
#boto==2.38.0

# Image manipulation
#django-versatileimagefield==1.2

# For asynchronous queuing
#django-rq==0.9.0

# Time zones support
pytz
django_extensions
django_cors_headers
